rtio: True async request chains
This removes the SYNC flag and now every submit to an iodev is expected
to be asynchronously done. The completion is then to be reported back
to the rtio context (and executor) through rtio_sqe_ok or rtio_sqe_err.

An almost complete mock up iodev based on ideas from Maureen Helm using
a k_timer to simulate asynchronicity is available.

A pair of optional semaphores has been added enabling potentially waiting
on submissions free or completions available to be possible. This needs
some more thought and cleanup though.

The tests now each have their own rtio context setup so the odd
resets are no longer needed.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>